requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asyncer==0.0.2
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chainlit==1.0.400
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
curl_cffi==0.6.2
dataclasses-json==0.5.14
Deprecated==1.2.14
duckduckgo_search==5.0
fastapi==0.108.0
fastapi-socketio==0.0.10
filelock==3.13.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==6.11.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-experimental==0.0.53
langchain-text-splitters==0.0.1
langsmith==0.1.23
Lazify==0.4.0
literalai==0.0.300
lxml==5.1.0
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numexpr==2.9.0
numpy==1.26.4
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
orjson==3.9.15
packaging==23.2
protobuf==4.25.3
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.11.1
PyYAML==6.0.1
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.32.0.post1
syncer==2.0.3
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
uptrace==1.22.0
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.20.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
